* development: (830 commits)
  [EDELIVERY-11449] confirmation window when deleting a certificate
  [EDELIVERY-11524]-Investigate and fix OWASP vulnerabilities
  EDELIVERY-11516 fix swagger client for earchive
  Update version to 5.2-SNAPSHOT
  [EDELIVERY-10990] fix tests
  [EDELIVERY-10990] address PR comments: domibus exception; log
  [EDELIVERY-10990] fix typo
  [EDELIVERY-10990] use constant
  [EDELIVERY-10990] adapt tests
  [EDELIVERY-10990] remove property
  [EDELIVERY-10990] rename method; add exception; add constant
  [EDELIVERY-11138] no need for decimal value
  [EDELIVERY-10990] adapt tests
  [EDELIVERY-10990] add and use a property to decide whether to keep the encryptedData section of a message
  [EDELIVERY-10990] extract the raw_message_xml earlier
  [EDELIVERY-11248] try adding an invalid certificate in test
  EDELIVERY-11450 - Add support for inspecting Content-Type header in MimeHeaders of a soap message
  [EDELIVERY-11248] add more verifications: stores are in sync
  [EDELIVERY-11248] reuse code in tests; add more change operations to the store and compare memory with file stores
  EDELIVERY-11450 - Add support for inspecting Content-Type header in MimeHeaders of a soap message
  ...

# Conflicts:
#	Core/Domibus-MSH-db/pom.xml
#	Tomcat/Domibus-MSH-tomcat/src/test/java/eu/domibus/AbstractIT.java

+++ Core/Domibus-MSH-angular/src/app/truststore/base-truststore.component.ts
@@ -177,7 +177,7 @@ export class BaseTruststoreComponent extends mix(BaseListComponent).with(ClientP
   }
 
   canAddCertificate() {
-    return this.storeExists  && !this.isBusy();
+    return this.storeExists && !this.isBusy();
   }
 
   canRemoveCertificate() {
@@ -194,6 +194,12 @@ export class BaseTruststoreComponent extends mix(BaseListComponent).with(ClientP
     if (!cert) {
       return;
     }
+
+    const confirm = await this.dialogsService.openYesNoDialog('Are you sure you want to delete the selected certificate?');
+    if (!confirm) {
+      return;
+    }
+    
     try {
       super.isLoading = true;
       let res = await this.truststoreService.removeCertificate(this.REMOVE_CERTIFICATE_URL, cert);
@@ -242,7 +248,7 @@ export class BaseTruststoreComponent extends mix(BaseListComponent).with(ClientP
   protected async checkModifiedOnDisk() {
     const isChanged = await this.http.get<boolean>(this.BASE_URL + '/changedOnDisk').toPromise();
     if (isChanged) {
-      const refresh = await this.dialogsService.openYesNoDialog('Store file on the disk is newer and has different content than the one loaded and used in Domibus. ' +
+      const refresh = await this.dialogsService.openYesNoDialog('The store file on the disk has different content than the one loaded and used in Domibus. ' +
         'Would you like to refresh?');
       if (refresh) {
         this.reloadStore();
